+++ Assets/RCore.SheetX/Editor/ExcelSheetHandler.cs
@@ -312,7 +312,6 @@ namespace RCore.SheetX.Editor
 		/// </summary>
 		public void ExportConstants()
 		{
-#if !SX_LOCALIZATION
 			if (string.IsNullOrEmpty(m_settings.constantsOutputFolder))
 			{
 				UnityEngine.Debug.LogError("Please setup the Constants Output Folder!");
@@ -356,12 +355,10 @@ namespace RCore.SheetX.Editor
 				}
 				m_settings.CreateFileConstants(builder.ToString(), "Constants");
 			}
-#endif
 		}
 
 		private void LoadSheetConstantsData(IWorkbook pWorkbook, string pSheetName)
 		{
-#if !SX_LOCALIZATION
 			var sheet = pWorkbook.GetSheet(pSheetName);
 			if (sheet == null || sheet.LastRowNum == 0)
 			{
@@ -421,12 +418,10 @@ namespace RCore.SheetX.Editor
 			}
 			constants.Sort();
 			BuildContentOfFileConstants(constants, pSheetName);
-#endif
 		}
 
 		private void BuildContentOfFileConstants(List<ConstantBuilder> constants, string constantsSheet)
 		{
-#if !SX_LOCALIZATION
 			var constantsSB = new StringBuilder("");
 			for (int i = 0; i < constants.Count; i++)
 			{
@@ -533,7 +528,6 @@ namespace RCore.SheetX.Editor
 			else
 				m_constantsBuilderDict.Add(constantsSheet, new StringBuilder());
 			m_constantsBuilderDict[constantsSheet].Append(constantsSB);
-#endif
 		}
 
 #endregion
@@ -545,7 +539,6 @@ namespace RCore.SheetX.Editor
 		/// </summary>
 		public void ExportLocalizations()
 		{
-#if !SX_NO_LOCALIZATION
 			if (string.IsNullOrEmpty(m_settings.constantsOutputFolder))
 			{
 				UnityEngine.Debug.LogError("Please setup the Constants Output Folder!");
@@ -613,7 +606,6 @@ namespace RCore.SheetX.Editor
 
 			//Create localization manager file
 			CreateLocalizationsManagerFile();
-#endif
 		}
 
 		private void LoadSheetLocalizationData(IWorkbook pWorkBook, string pSheetName)
@@ -712,7 +704,6 @@ namespace RCore.SheetX.Editor
 
 		private void CreateLocalizationFile(List<string> pIdsString, Dictionary<string, List<string>> pLanguageTextDict, string pFileName)
 		{
-#if !SX_NO_LOCALIZATION
 			if (pLanguageTextDict.Count == 0 || pLanguageTextDict.Count == 0)
 				return;
 
@@ -823,12 +814,10 @@ namespace RCore.SheetX.Editor
 			fileContent = SheetXHelper.AddNamespace(fileContent, m_settings.@namespace);
 			SheetXHelper.WriteFile(m_settings.constantsOutputFolder, $"{pFileName}Text.cs", fileContent);
 			UnityEngine.Debug.Log($"Exported {pFileName}Text.cs!");
-#endif
 		}
 
 		private void CreateLocalizationsManagerFile()
 		{
-#if !SX_NO_LOCALIZATION
 			//Create language character sets
 			if (m_langCharSets != null && m_langCharSets.Count > 0)
 			{
@@ -960,7 +949,6 @@ namespace RCore.SheetX.Editor
 				SheetXHelper.WriteFile(m_settings.constantsOutputFolder, "LocalizationsManager.cs", fileContent);
 				UnityEngine.Debug.Log($"Exported LocalizationsManager.cs!");
 			}
-#endif
 		}
 
 #endregion
@@ -969,7 +957,6 @@ namespace RCore.SheetX.Editor
 
 		public void ExportJson()
 		{
-#if !SX_LOCALIZATION
 			if (string.IsNullOrEmpty(m_settings.jsonOutputFolder))
 			{
 				UnityEngine.Debug.LogError("Please setup the Json Output folder!");
@@ -1021,23 +1008,18 @@ namespace RCore.SheetX.Editor
 				else
 					UnityEngine.Debug.Log($"Exported Json data to {mergedFileName}.txt.");
 			}
-#endif
 		}
 
 		private string ConvertSheetToJson(IWorkbook pWorkBook, string pSheetName, string pFileName, bool pEncrypt, bool pWriteFile)
 		{
-#if !SX_LOCALIZATION
 			var fieldValueTypes = SheetXHelper.GetFieldValueTypes(pWorkBook, pSheetName);
 			if (fieldValueTypes == null)
 				return "{}";
 			return ConvertSheetToJson(pWorkBook, pSheetName, pFileName, fieldValueTypes, pEncrypt, pWriteFile);
-#endif
-			return "{}";
 		}
 
 		private string ConvertSheetToJson(IWorkbook pWorkBook, string pSheetName, string pOutputFile, List<FieldValueType> pFieldValueTypes, bool pEncrypt, bool pAutoWriteFile)
 		{
-#if !SX_LOCALIZATION
 
 			var sheet = pWorkBook.GetSheet(pSheetName);
 			if (sheet == null || sheet.LastRowNum == 0)
@@ -1580,8 +1562,6 @@ namespace RCore.SheetX.Editor
 					UnityEngine.Debug.Log($"Exported Json data to {pOutputFile}.txt.");
 			}
 			return finalContent;
-#endif
-			return "{}";
 		}
 
 #endregion
@@ -1596,7 +1576,6 @@ namespace RCore.SheetX.Editor
 
 		public void ExportAllFiles()
 		{
-#if !SX_LITE
 			m_idsBuilderDict = new Dictionary<string, StringBuilder>();
 			m_constantsBuilderDict = new Dictionary<string, StringBuilder>();
 			m_localizationsDict = new Dictionary<string, LocalizationBuilder>();
@@ -1687,7 +1666,6 @@ namespace RCore.SheetX.Editor
 					if (m_constantsBuilderDict.ContainsKey(sheet.name) && m_settings.separateConstants)
 						m_settings.CreateFileConstants(m_constantsBuilderDict[sheet.name].ToString(), sheet.name);
 				}
-#if !SX_NO_LOCALIZATION
 				//Load and write localizations
 				foreach (var sheet in file.sheets)
 				{
@@ -1703,7 +1681,6 @@ namespace RCore.SheetX.Editor
 						m_localizedSheetsExported.Add(sheet.name);
 					}
 				}
-#endif
 			}
 
 			//Create file contain all IDs
@@ -1738,7 +1715,6 @@ namespace RCore.SheetX.Editor
 				m_settings.CreateFileConstants(builder.ToString(), "Constants");
 			}
 
-#if !SX_NO_LOCALIZATION
 			//Create file contain all Localizations
 			if (!m_settings.separateLocalizations)
 			{
@@ -1758,13 +1734,11 @@ namespace RCore.SheetX.Editor
 				CreateLocalizationFile(localizationBuilder.idsString, localizationBuilder.languageTextDict, "Localization");
 				m_localizedSheetsExported.Add("Localization");
 			}
-#endif
 
 			//Create localization manager file
 			CreateLocalizationsManagerFile();
 
 			UnityEngine.Debug.Log("Done!");
-#endif
 		}
 
 		private bool CheckExistedId(string pKey)